fix(condo): DOMA-10399 move date range picker to separate component
