+++ apps/condo/domains/meter/components/TabContent/MeterReading.tsx
@@ -6,14 +6,12 @@ import { Meter as MeterType, MeterResource as MeterResourceType } from '@app/con
 import { Col, Row, RowProps } from 'antd'
 import { CheckboxChangeEvent } from 'antd/lib/checkbox/Checkbox'
 import { TableRowSelection } from 'antd/lib/table/interface'
-import dayjs, { Dayjs } from 'dayjs'
 import chunk from 'lodash/chunk'
 import get from 'lodash/get'
 import uniqBy from 'lodash/uniqBy'
 import { useRouter } from 'next/router'
-import React, { CSSProperties, useCallback, useEffect, useMemo, useState } from 'react'
+import React, { CSSProperties, useCallback, useMemo, useState } from 'react'
 
-import { useDeepCompareEffect } from '@open-condo/codegen/utils/useDeepCompareEffect'
 import { PlusCircle, Search } from '@open-condo/icons'
 import { useAuth } from '@open-condo/next/auth'
 import { useIntl } from '@open-condo/next/intl'
@@ -25,7 +23,6 @@ import { TablePageContent } from '@condo/domains/common/components/containers/Ba
 import { DeleteButtonWithConfirmModal } from '@condo/domains/common/components/DeleteButtonWithConfirmModal'
 import { EmptyListContent } from '@condo/domains/common/components/EmptyListContent'
 import { Loader } from '@condo/domains/common/components/Loader'
-import DateRangePicker from '@condo/domains/common/components/Pickers/DateRangePicker'
 import { DEFAULT_PAGE_SIZE, Table } from '@condo/domains/common/components/Table/Index'
 import { TableFiltersContainer } from '@condo/domains/common/components/TableFiltersContainer'
 import { useTracking } from '@condo/domains/common/components/TrackingContext'
@@ -37,9 +34,9 @@ import { useSearch } from '@condo/domains/common/hooks/useSearch'
 import { FiltersMeta } from '@condo/domains/common/utils/filters.utils'
 import { getPageIndexFromOffset, parseQuery } from '@condo/domains/common/utils/tables.utils'
 import { MetersImportWrapper } from '@condo/domains/meter/components/Import/Index'
+import { MeterReadingDatePicker } from '@condo/domains/meter/components/MeterReadingDatePicker'
 import ActionBarForSingleMeter from '@condo/domains/meter/components/Meters/ActionBarForSingleMeter'
 import UpdateMeterReadingModal from '@condo/domains/meter/components/Meters/UpdateMeterReadingModal'
-import { EXPORT_METER_READINGS_MONTHS_LIMIT } from '@condo/domains/meter/constants/constants'
 import { useMeterReadingExportToExcelTask } from '@condo/domains/meter/hooks/useMeterReadingExportToExcelTask'
 import { useTableColumns } from '@condo/domains/meter/hooks/useTableColumns'
 import {
@@ -56,7 +53,6 @@ const METERS_PAGE_CONTENT_ROW_GUTTERS: RowProps['gutter'] = [0, 40]
 const FILTERS_CONTAINER_GUTTER: RowProps['gutter'] = [16, 16]
 const RESET_FILTERS_BUTTON_STYLE: CSSProperties = { paddingLeft: 0 }
 const QUICK_FILTERS_COL_STYLE: CSSProperties = { alignSelf: 'center' }
-const FULL_WIDTH_DATE_RANGE_STYLE: CSSProperties = { width: '100%' }
 
 const SORTABLE_PROPERTIES = ['date', 'clientName', 'source']
 
@@ -90,8 +86,6 @@ const MeterReadingsTableContent: React.FC<MetersTableContentProps> = ({
     const intl = useIntl()
     const CreateMeterReadingsButtonLabel = intl.formatMessage({ id: 'pages.condo.meter.index.CreateMeterReadingsButtonLabel' })
     const SearchPlaceholder = intl.formatMessage({ id: 'filters.FullSearch' })
-    const StartDateMessage = intl.formatMessage({ id: 'pages.condo.meter.StartDate' })
-    const EndDateMessage = intl.formatMessage({ id: 'pages.condo.meter.EndDate' })
     const CancelSelectedReadingsMessage = intl.formatMessage({ id: 'global.cancelSelection' })
     const CountSelectedReadingsMessage = intl.formatMessage({ id: 'pages.condo.meter.MeterReading.CountSelectedReadings' })
     const DeleteMeterReadingsMessage = intl.formatMessage({ id: 'pages.condo.meter.MeterReading.DeleteMeterReadings' })
@@ -112,7 +106,7 @@ const MeterReadingsTableContent: React.FC<MetersTableContentProps> = ({
     const sortBy = useMemo(() => sortersToSortBy(sorters) as SortMeterReadingsBy[], [sorters, sortersToSortBy])
     const { getTrackingWrappedCallback } = useTracking()
 
-    const [dateRange, setDateRange] = useDateRangeSearch('date')
+    const [dateRange] = useDateRangeSearch('date')
     const dateFilterValue = dateRange || null
     const dateFilter = dateFilterValue ? dateFilterValue.map(el => el.toISOString()) : null
     const nextVerificationDate = get(meter, 'nextVerificationDate')
@@ -120,10 +114,6 @@ const MeterReadingsTableContent: React.FC<MetersTableContentProps> = ({
 
     const [isShowUpdateReadingModal, setIsShowUpdateReadingModal] = useState(false)
     const [chosenMeterReadingId, setChosenMeterReadingId] = useState<string>(null)
-
-    const handleDateChange = useCallback((value) => {
-        setDateRange(value)
-    }, [setDateRange])
 
     const searchMeterReadingsQuery = useMemo(() => ({
         ...filtersToWhere({ ... !meter ? { date: dateFilter } : {}, ...filters }),
@@ -250,35 +240,6 @@ const MeterReadingsTableContent: React.FC<MetersTableContentProps> = ({
     const handleCreateMeterReadings = useCallback(() => router.push(`/meter/create?tab=${METER_TAB_TYPES.meterReading}`), [router])
     const handleCloseUpdateReadingModal = useCallback(() => setIsShowUpdateReadingModal(false), [])
 
-    const [selectedDates, setSelectedDates] = useState<[Dayjs, Dayjs]>()
-
-    const onCalendarChange = useCallback(dates => setSelectedDates(dates), [])
-
-    const disabledDate = useCallback((current) => {
-        if (current > dayjs()) return true
-        if (!selectedDates) return false
-
-        const startDate = selectedDates[0]
-        const endDate = selectedDates[1]
-        const tooLate = startDate && current.diff(startDate, 'months', true) > EXPORT_METER_READINGS_MONTHS_LIMIT
-        const tooEarly = endDate && endDate.diff(current, 'months', true) > EXPORT_METER_READINGS_MONTHS_LIMIT
-
-        return !!tooEarly || !!tooLate
-    }, [selectedDates])
-
-    const onOpenChange = useCallback((open: boolean) => {
-        if (open || !selectedDates) {
-            return
-        }
-        if (!selectedDates[0] || !selectedDates[1]) {
-            setSelectedDates(null)
-        }
-    }, [selectedDates])
-
-    useDeepCompareEffect(() => {
-        setSelectedDates(dateRange)
-    }, [dateRange])
-
     return (
         <>
             <Row
@@ -301,15 +262,7 @@ const MeterReadingsTableContent: React.FC<MetersTableContentProps> = ({
                                 <Col xs={24} sm={24} md={16}>
                                     <Row justify='start' gutter={FILTERS_CONTAINER_GUTTER} style={{ flexWrap: 'nowrap' }}>
                                         <Col style={QUICK_FILTERS_COL_STYLE} xs={24} sm={11}>
-                                            <DateRangePicker
-                                                value={dateRange}
-                                                onChange={handleDateChange}
-                                                placeholder={[StartDateMessage, EndDateMessage]}
-                                                disabledDate={disabledDate}
-                                                style={FULL_WIDTH_DATE_RANGE_STYLE}
-                                                onCalendarChange={onCalendarChange}
-                                                onOpenChange={onOpenChange}
-                                            />
+                                            <MeterReadingDatePicker/>
                                         </Col>
                                     </Row>
                                 </Col>
